Generate an example data frame of clinical codes for diabetes

Data frames of this format may be used with extract_phenotypes. Use validate_clinical_codes to check whether a clinical codes data frame meets all requirements.

Usage

example_clinical_codes()

Value

A data frame of selected clinical codes (non-exhaustive) for diabetes.

See also

Examples

example_clinical_codes()
#> # A tibble: 8 × 6
#>   disease  description                           category code_type code  author
#>   <chr>    <chr>                                 <chr>    <chr>     <chr> <chr> 
#> 1 Diabetes diabetes                              Diabete… data_cod… 1220  ukbwr 
#> 2 Diabetes gestational diabetes                  Gestati… data_cod… 1221  ukbwr 
#> 3 Diabetes type 1 diabetes                       Type 1 … data_cod… 1222  ukbwr 
#> 4 Diabetes type 2 diabetes                       Type 2 … data_cod… 1223  ukbwr 
#> 5 Diabetes Type 1 diabetes mellitus              Type 1 … icd10     E10   ukbwr 
#> 6 Diabetes Type 2 diabetes mellitus              Type 2 … icd10     E11   ukbwr 
#> 7 Diabetes Insulin dependent diabetes mellitus   Type 1 … read2     C108. ukbwr 
#> 8 Diabetes Non-insulin dependent diabetes melli… Type 2 … read2     C109. ukbwr