Get the earliest record date for multiple phenotypes

Builds on extract_first_or_last_clinical_event, extracting either the first or last recorded clinical event for multiple phenotypes.

extract_first_or_last_clinical_event_multi(
  df,
  clinical_codes_df,
  min_max = "min",
  prefix = NULL
)

Arguments

df
Either a dataframe or a tbl object generated by one of the get_XXX_diagnoses functions

clinical_codes_df
data frame. Must match the format as per generate_self_reported_diabetes_codes_df.

min_max
can be either "min" or "max". Extracts either the earliest or latest date that on of the specified codes appears.

prefix
character. Optionally add a prefix to column names.

Value

Data frame with an "eid" column, and "event_min/max_indicator" and "event_min/max_date" columns for each phenotype in the 'category' column of clinical_codes_df.

See also