Extract phenotypes from clinical events data

Filters a clinical events table created by tidy_clinical_events for a set or sets of specified clinical codes that represent one or more phenotypes. By default, the earliest date that any clinical code appears in an individual participant's record is extracted. See also the 'Clinical events' vignette on the ukbwranglr package website.

extract_phenotypes(
  clinical_events,
  clinical_codes,
  data_sources = NULL,
  min_max = "min",
  colnames_prefix = NULL,
  labels_prefix = NULL,
  workers = NULL,
  keep_all = FALSE
)

Arguments

clinical_events

A long format data frame created by tidy_clinical_events, tidy_gp_clinical, tidy_gp_scripts or make_clinical_events_db. This can also be a tbl_dbi object.

clinical_codes

data frame. Must match the format as per example_clinical_codes.

data_sources

A character vector of clinical events sources in clinical_events to extract phenotypes from. Use clinical_events_sources (source column) for a list of valid values.

min_max

Choose either "min" or "max" to extract either the earliest or latest date that one of the codes appears in clinical_codes appears.

colnames_prefix

Optionally add a prefix to column names.

labels_prefix

Optionally add a prefix to variable labels.

workers

The number of processes to run in parallel when clinical_codes includes multiple diseases. This is used to set the number of workers in plan with strategy = multisession, which is passed on to future_map. If NULL, then no processing is performed sequentially. Default value is NULL.

keep_all

If TRUE, keep all matching codes for each eid in a nested data frame column named 'data'.

Value

A named list of data frames, one for each disease. Each data frame has an "eid" column, and "event_min/max_indicator" and "event_min/max_date" columns for each phenotype in the 'category' column of clinical_codes for that disease. If keep_all is TRUE, then there will also be additional nested data frame column called 'data'.

See also

Examples

library(magrittr)

# dummy clinical events data frame
dummy_ukb_data_dict <- get_ukb_dummy("dummy_Data_Dictionary_Showcase.tsv")
dummy_ukb_codings <- get_ukb_dummy("dummy_Codings.tsv")

dummy_clinical_events <- read_ukb(
  path = get_ukb_dummy("dummy_ukb_main.tsv", path_only = TRUE),
  ukb_data_dict = dummy_ukb_data_dict,
  ukb_codings = dummy_ukb_codings
) %>%
  tidy_clinical_events(
    ukb_data_dict = dummy_ukb_data_dict,
    ukb_codings = dummy_ukb_codings
  ) %>%
  dplyr::bind_rows()
#> Creating data dictionary
#> STEP 1 of 3
#> Reading data into R
#> STEP 2 of 3
#> Renaming with descriptive column names
#> STEP 3 of 3
#> Applying variable and value labels
#> Labelling dataset
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for primary_death_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for secondary_death_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_medication
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_non_cancer
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_non_cancer_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_cancer
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_operation
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for cancer_register_icd9
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for cancer_register_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_icd9
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_opcs3
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_opcs4
#> Time taken: 0 minutes, 0 seconds.

head(dummy_clinical_events)
#>    eid source index code       date
#> 1:   1 f40001   0_0 X095 1917-10-08
#> 2:   2 f40001   0_0 A162 1955-02-11
#> 3:   1 f40001   1_0 X095 1910-02-19
#> 4:   2 f40001   1_0 A162 1965-08-08
#> 5:   1 f40002   0_0 W192 1917-10-08
#> 6:   2 f40002   0_0 V374 1955-02-11

# dummy clinical code list
example_clinical_codes()
#> # A tibble: 8 × 6
#>   disease  description                           category code_type code  author
#>   <chr>    <chr>                                 <chr>    <chr>     <chr> <chr> 
#> 1 Diabetes diabetes                              Diabete… data_cod… 1220  ukbwr 
#> 2 Diabetes gestational diabetes                  Gestati… data_cod… 1221  ukbwr 
#> 3 Diabetes type 1 diabetes                       Type 1 … data_cod… 1222  ukbwr 
#> 4 Diabetes type 2 diabetes                       Type 2 … data_cod… 1223  ukbwr 
#> 5 Diabetes Type 1 diabetes mellitus              Type 1 … icd10     E10   ukbwr 
#> 6 Diabetes Type 2 diabetes mellitus              Type 2 … icd10     E11   ukbwr 
#> 7 Diabetes Insulin dependent diabetes mellitus   Type 1 … read2     C108. ukbwr 
#> 8 Diabetes Non-insulin dependent diabetes melli… Type 2 … read2     C109. ukbwr 

# Filter for participants with matching clinical codes,
# by default only the earliest date is extracted
cases <- extract_phenotypes(
  clinical_events = dummy_clinical_events,
  clinical_codes = example_clinical_codes()
)
#> 
#> ***PROCESSING DISEASE 1 OF 1***
#> Time taken: 0 minutes, 0 seconds.
#> 
#> Extracting clinical events for disease: Diabetes
#> Filtering clinical_events for all codes in clinical_codes
#> EXTRACTING EVENT DATES FOR PHENOTYPES
#> Extracting event dates for DIABETES_UKBWR (Phenotype 1 of 5)
#> Time taken: 0 minutes, 1 seconds.
#> Extracting event dates for diabetes_unspecified_ukbwr (Phenotype 2 of 5)
#> Time taken: 0 minutes, 1 seconds.
#> Extracting event dates for gestational_diabetes_ukbwr (Phenotype 3 of 5)
#> Time taken: 0 minutes, 2 seconds.
#> Extracting event dates for type_1_dm_ukbwr (Phenotype 4 of 5)
#> Time taken: 0 minutes, 2 seconds.
#> Extracting event dates for type_2_dm_ukbwr (Phenotype 5 of 5)
#> Time taken: 0 minutes, 3 seconds.
#> COMPLETE!
#> Time taken: 0 minutes, 3 seconds.

# returns a named list of data frames, one for each category in
# lower case, and one for the overall disease in capitals
cases
#> $Diabetes
#> $Diabetes$DIABETES_UKBWR
#> # A tibble: 2 × 3
#>     eid DIABETES_UKBWR_min_date DIABETES_UKBWR_indicator
#>   <int> <chr>                   <chr>                   
#> 1     1 1910-02-19              Yes                     
#> 2     2 1939-02-16              Yes                     
#> 
#> $Diabetes$diabetes_unspecified_ukbwr
#> NULL
#> 
#> $Diabetes$gestational_diabetes_ukbwr
#> NULL
#> 
#> $Diabetes$type_1_dm_ukbwr
#> # A tibble: 1 × 3
#>     eid type_1_dm_ukbwr_min_date type_1_dm_ukbwr_indicator
#>   <int> <chr>                    <chr>                    
#> 1     1 1910-02-19               Yes                      
#> 
#> $Diabetes$type_2_dm_ukbwr
#> # A tibble: 2 × 3
#>     eid type_2_dm_ukbwr_min_date type_2_dm_ukbwr_indicator
#>   <int> <chr>                    <chr>                    
#> 1     1 2003-02-25               Yes                      
#> 2     2 1939-02-16               Yes                      
#> 
#>