Write a file to a database

Writes a file in chunks to a SQLite database, with or without pre-processing.

file_to_sqlite_db(
  file,
  table,
  db_path = "ukb.db",
  chunk_size = 10000,
  delim = "\t",
  col_types = readr::cols(.default = "c"),
  verbose = TRUE,
  append = FALSE,
  data_processing_function = NULL,
  quote = "\"",
  escape_backslash = FALSE,
  escape_double = TRUE,
  col_names = TRUE,
  locale = readr::default_locale(),
  na = c("", "NA"),
  quoted_na = TRUE,
  comment = "",
  trim_ws = FALSE,
  skip = 0,
  guess_max = min(1000, chunk_size),
  progress = readr::show_progress(),
  skip_empty_rows = TRUE,
  ...
)

Arguments

file
Character. The path to the file to be written to db_path.

table
Character. Name of table to write to.

db_path
Character. Path to a SQLite database. A new database will be created if this does not already exist. Default is "ukb.db".

chunk_size
The number of rows to include in each chunk. Default is 10000.

delim
Single character used to separate fields within a record.

col_types
Default is for all columns to be type character.

verbose
Print time taken after each chunk has been written to database.

append
Append to table if already exists. Default is FALSE.

data_processing_function
A function to be applied to each chunk before writing to database. Default is `NULL`.

quote
Single character used to quote strings.

escape_backslash
Does the file use backslashes to escape special characters? This is more general than escape_double as backslashes can be used to escape the delimiter character, the quote character, or to add special characters like \\n.

escape_double
Does the file escape quotes by doubling them? i.e. If this option is TRUE, the value """" represents a single quote, \".

col_names
Either TRUE, FALSE or a character vector of column names.

If TRUE, the first row of the input will be used as the column names, and will not be included in the data frame. If FALSE, column names will be generated automatically: X1, X2, X3 etc.

If col_names is a character vector, the values will be used as the names of the columns, and the first row of the input will be read into the first row of the output data frame.

Missing (NA) column names will generate a warning, and be filled in with dummy names X1, X2 etc. Duplicate column names will generate a warning and be made unique with a numeric suffix.

locale
The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

na
Character vector of strings to interpret as missing values. Set this option to character() to indicate no missing values.

quoted_na
Should missing values inside quotes be treated as missing values (the default) or strings.

comment
A string used to identify comments. Any text after the comment characters will be silently ignored.

trim_ws
Should leading and trailing whitespace be trimmed from each field before parsing it?

skip
Number of lines to skip before reading data.

guess_max
Maximum number of records to use for guessing column types.

progress
Display a progress bar? By default it will only display in an interactive session and not while knitting a document. The display is updated every 50,000 values and will only display if estimated reading time is 5 seconds or more. The automatic progress bar can be disabled by setting option readr.show_progress to FALSE.

skip_empty_rows
Should blank rows be ignored altogether? i.e. If this option is TRUE then blank rows will not be represented at all. If it is FALSE then they will be represented by NA values in all the columns.

...
additional parameters passed on to data_processing_function.

Value

Returns a database connection object invisibly.

Details

This function is designed to be used with large files that may not fit into memory on a personal computer. The file is written to a table in a SQLite database in chunks with the option to apply a 'pre-processing' function to each chunk (e.g. reshape the data).

Under the hood

Relies on the DBI and RSQLite packages to create and write to a SQLite database, and read_delim_chunked to read the input file in chunks.

See also