Get all diagnostic codes from multiple data sources

Extract diagnostic codes from multiple sources

get_all_diagnostic_codes_multi(
  function_list = list(get_self_report_non_cancer_diagnoses_icd10,
    get_hes_icd9_diagnoses, get_hes_icd10_diagnoses, get_death_data_icd10_diagnoses,
    get_cancer_register_icd9_diagnoses, get_cancer_register_icd10_diagnoses),
  ukb_pheno,
  data_dict,
  ukb_codings,
  drop_source_col = TRUE
)

Arguments

function_list
A list of `get_XXXX_diagnoses` functions.

ukb_pheno
a UK Biobank phenotype dataset

data_dict
a data dictionary for ukb_pheno

ukb_codings
The path to a tab-delimited UKB codings dictionary. By default, this is downloaded directly from the UKB website.

drop_source_col
Remove `source_col` column, which records which column in the main dataset the value in `code` column came from. Default is `TRUE`.

Value

Dataframe

Details

Loops through a list of functions (`function_list`) from the 'get all diagnostic codes' family and combines the results into a single dataframe.

See also