Dummy UK Biobank data

Either read a dummy UK Biobank dataset into R or return the file path only.

get_ukb_dummy(file_name, path_only = FALSE)

Arguments

file_name

Name of dummy dataset file.

path_only

If `TRUE`, return the file path to the dummy dataset file, otherwise if `FALSE` (default), read the dummy dataset into R.

Value

A data frame if `path_only` is `FALSE` (default) or a string if `path_only` is `TRUE`.

Details

The following dummy datasets are included with this package:

- `dummy_Data_Dictionary_Showcase.tsv`: A subset of fields from the UK Biobank data dictionary (full version available from the UK Biobank [data showcase website](https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide)).

- `dummy_Codings.tsv`: A subset of UK Biobank data codings (full version available from the UK Biobank [data showcase website](https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide)).

- `dummy_ukb_main.tsv`: A dummy main UK Biobank dataset. May be read into R with [read_ukb()]. Tidy xlinical events fields with [tidy_clinical_events()].

- `dummy_gp_clinical.txt`: A dummy UK Biobank primary care clinical event records dataset.

- `dummy_gp_scripts.txt`: A dummy UK Biobank primary care prescription records dataset.

Examples

library(magrittr)

# available dummy datasets
dummy_datasets <- c(
  "dummy_Data_Dictionary_Showcase.tsv",
  "dummy_Codings.tsv",
  "dummy_ukb_main.tsv",
  "dummy_gp_clinical.txt",
  "dummy_gp_scripts.txt"
)

# file paths
dummy_datasets %>%
  purrr::map_chr(get_ukb_dummy, path_only = TRUE)
#> [1] "/Users/runner/work/ukbwranglr/ukbwranglr/renv/library/R-4.2/x86_64-apple-darwin17.0/ukbwranglr/extdata/dummy_Data_Dictionary_Showcase.tsv"
#> [2] "/Users/runner/work/ukbwranglr/ukbwranglr/renv/library/R-4.2/x86_64-apple-darwin17.0/ukbwranglr/extdata/dummy_Codings.tsv"                 
#> [3] "/Users/runner/work/ukbwranglr/ukbwranglr/renv/library/R-4.2/x86_64-apple-darwin17.0/ukbwranglr/extdata/dummy_ukb_main.tsv"                
#> [4] "/Users/runner/work/ukbwranglr/ukbwranglr/renv/library/R-4.2/x86_64-apple-darwin17.0/ukbwranglr/extdata/dummy_gp_clinical.txt"             
#> [5] "/Users/runner/work/ukbwranglr/ukbwranglr/renv/library/R-4.2/x86_64-apple-darwin17.0/ukbwranglr/extdata/dummy_gp_scripts.txt"              

# read into R
dummy_datasets %>%
  purrr::set_names() %>%
  purrr::map(get_ukb_dummy, path_only = FALSE) %>%
  purrr::map(tibble::as_tibble)
#> $dummy_Data_Dictionary_Showcase.tsv
#> # A tibble: 28 × 17
#>    Path      Category FieldID Field Participants Items Stability ValueType Units
#>    <chr>     <chr>    <chr>   <chr> <chr>        <chr> <chr>     <chr>     <chr>
#>  1 Populati… 100094   31      Sex   502413       5024… Complete  Categori… NA   
#>  2 Populati… 100094   34      Year… 502413       5024… Complete  Integer   years
#>  3 Populati… 100094   52      Mont… 502413       5024… Complete  Categori… NA   
#>  4 Assessme… 100024   53      Date… 502414       5795… Complete  Date      NA   
#>  5 Assessme… 100011   4080    Syst… 475231       1061… Complete  Integer   mmHg 
#>  6 Assessme… 100074   20001   Canc… 45950        54022 Complete  Categori… NA   
#>  7 Assessme… 100074   20002   Non-… 386743       1145… Complete  Categori… NA   
#>  8 Assessme… 100075   20003   Trea… 373347       1389… Complete  Categori… NA   
#>  9 Assessme… 100076   20004   Oper… 399178       1003… Complete  Categori… NA   
#> 10 Assessme… 100074   20006   Inte… 45950        54022 Complete  Continuo… years
#> # … with 18 more rows, and 8 more variables: ItemType <chr>, Strata <chr>,
#> #   Sexed <chr>, Instances <chr>, Array <chr>, Coding <chr>, Notes <chr>,
#> #   Link <chr>
#> 
#> $dummy_Codings.tsv
#> # A tibble: 425 × 3
#>    Coding Value Meaning 
#>    <chr>  <chr> <chr>   
#>  1 8      1     January 
#>  2 8      10    October 
#>  3 8      11    November
#>  4 8      12    December
#>  5 8      2     February
#>  6 8      3     March   
#>  7 8      4     April   
#>  8 8      5     May     
#>  9 8      6     June    
#> 10 8      7     July    
#> # … with 415 more rows
#> 
#> $dummy_ukb_main.tsv
#> # A tibble: 10 × 71
#>      eid `31-0.0` `34-0.0` `52-0.0` `21000-0.0` `21000-1.0` `21000-2.0`
#>    <int> <chr>    <chr>    <chr>    <chr>       <chr>       <chr>      
#>  1     1 0        1952     8        -1          2           3003       
#>  2     2 0        1946     3        -3          2001        3004       
#>  3     3 1        1951     4        1           2002        -1         
#>  4     4 0        1956     9        1001        2003        4001       
#>  5     5 NA       NA       4        1002        2004        4002       
#>  6     6 1        1948     2        1003        3           4003       
#>  7     7 0        1949     12       NA          3001        5          
#>  8     8 1        1956     10       NA          5           NA         
#>  9     9 0        1962     4        4001        NA          NA         
#> 10    10 1        1953     2        4001        NA          NA         
#> # … with 64 more variables: `21001-0.0` <chr>, `21001-1.0` <chr>,
#> #   `21001-2.0` <chr>, `4080-0.0` <chr>, `4080-0.1` <chr>, `4080-0.2` <chr>,
#> #   `4080-0.3` <chr>, `4080-1.0` <chr>, `4080-1.1` <chr>, `4080-1.2` <chr>,
#> #   `4080-1.3` <chr>, `20001-0.0` <chr>, `20001-0.3` <chr>, `20001-2.0` <chr>,
#> #   `20001-2.3` <chr>, `20002-0.0` <chr>, `20002-0.3` <chr>, `20002-2.0` <chr>,
#> #   `20002-2.3` <chr>, `20006-0.0` <chr>, `20006-0.3` <chr>, `20006-2.0` <chr>,
#> #   `20006-2.3` <chr>, `20008-0.0` <chr>, `20008-0.3` <chr>, …
#> 
#> $dummy_gp_clinical.txt
#> # A tibble: 12 × 8
#>      eid data_provider event_dt   read_2 read_3 value1 value2 value3
#>    <int> <chr>         <chr>      <chr>  <chr>  <chr>  <chr>  <chr> 
#>  1     1 1             03/03/1903 C      NA     1      2      3     
#>  2     1 4             01/01/1901 A      NA     1      2      3     
#>  3     1 3             07/07/2037 E      NA     1      2      3     
#>  4     3 1             07/07/2037 E      NA     1      2      3     
#>  5     4 2             01/02/1999 J      NA     1      2      3     
#>  6     8 1             01/02/1999 NA     G      1      2      3     
#>  7     1 1             01/10/1990 C108.  NA     NA     NA     NA    
#>  8     2 2             02/10/1990 C109.  NA     NA     NA     NA    
#>  9     1 3             03/10/1990 NA     X40J4  NA     NA     NA    
#> 10     2 4             04/10/1990 NA     X40J5  NA     NA     NA    
#> 11     1 1             03/10/1990 NA     C108.  NA     NA     NA    
#> 12     2 2             04/10/1990 NA     C109.  NA     NA     NA    
#> 
#> $dummy_gp_scripts.txt
#> # A tibble: 6 × 8
#>     eid data_provider issue_date read_2 bnf_code     dmd_code drug_name quantity
#>   <int> <chr>         <chr>      <chr>  <chr>        <chr>    <chr>     <chr>   
#> 1     1 1             03/03/1903 bxi300 NA           1        drug2     50      
#> 2     1 4             01/01/1901 bxi3   NA           NA       NA        NA      
#> 3     1 3             07/07/2037 NA     02.02.01.00… NA       drug2     30      
#> 4     3 1             07/07/2037 bd3j00 NA           1        drug2     30      
#> 5     4 2             01/02/1999 bd3j   02020100     NA       drug2     30      
#> 6     8 1             01/02/1999 NA     NA           1        2         30      
#>