Pre-processing function - reformat UK Biobank primary care clinical events data

To be used with file_to_sqlite_db. Reformats the UK Biobank primary care clinical events dataset to match the output format for get_all_diagnostic_codes_multi.

gp_clinical_to_sqlite_db(df, remove_special_dates = TRUE)

Arguments

df
A dataframe

remove_special_dates
Logical. Removes special date values if requested (see resource 591 for details). Default is TRUE.

Value

data frame

Under the hood

The UK Biobank primary care data lists read codes in separate columns, one for Read2 and one for Read3. This function reshapes the data to long format so that all codes are in a single column. Note that the 3 free-text 'value' columns are dropped. By default, special date values are set to NA.

See also