Read a UK Biobank main dataset file

Reads a UK Biobank main dataset file into R using either fread or read_dta. Optionally renames variables with descriptive names and applies STATA-style variable and value labels using the haven package.

read_ukb(
  path,
  delim = "\t",
  data_dict = NULL,
  ukb_data_dict = get_ukb_data_dict(),
  ukb_codings = get_ukb_codings(),
  descriptive_colnames = TRUE,
  labelled = TRUE,
  max_n_labels = 30,
  na.strings = c("", "NA"),
  nrows = Inf,
  ...
)

Arguments

path
The path to a UK Biobank main dataset file.

delim
Delimiter for the UK Biobank main dataset file. Default is \t

data_dict
A data dictionary specific to the UKB main dataset file, generated by make_data_dict. To load only a selection of columns, supply a filtered copy of this data dictionary containing only the required variables. If NULL (default) then all fields will be read.

ukb_data_dict
The UKB data dictionary (available online at the UK Biobank data showcase. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

ukb_codings
The UKB codings dictionary. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

descriptive_colnames
If TRUE, rename columns with longer descriptive names derived from the UK Biobank's data dictionary 'Field' column.

labelled
If TRUE, apply variable and value labels using labelled.

max_n_labels
Coded variables with associated value labels up to (and equal to) this threshold will be labelled using. If NULL, then all value labels will be applied. Default value is 30. labelled if labelled is TRUE.

na.strings
A character vector of strings which are to be interpreted as NA values. By default, ",," for columns of all types, including type character is read as NA for consistency. ,"", is unambiguous and read as an empty string. To read ,NA, as NA, set na.strings="NA". To read ,, as blank string "", set na.strings=NULL. When they occur in the file, the strings in na.strings should not appear quoted since that is how the string literal ,"NA", is distinguished from ,NA,, for example, when na.strings="NA".

nrows
The maximum number of rows to read. Unlike read.table, you do not need to set this to an estimate of the number of rows in the file for better speed because that is already automatically determined by fread almost instantly using the large sample of lines. nrows=0 returns the column names and typed empty columns determined by the large sample; useful for a dry run of a large file or to quickly check format consistency of a set of files before starting to read any of them.

...
Additional parameters are passed on to either fread or read_dta

Value

A UK Biobank phenotype dataset as a data table with human-readable variables labels and data values.

Details

Note that na.strings is not recognised by read_dta. Reading in a STATA file may therefore require careful checking for empty strings that ned converting to NA.