Reformat ICD-10 codes

The lookup sheet in UKB resource 592 for ICD-10 ("icd10_lkp") has a column called "ALT_CODE", which is an alternative format for ICD-10 codes. This is the format used in the mapping sheets for this resource, as well as in Field ID 41270. This function converts from one format to the other.

reformat_icd10_codes(
  icd10_codes,
  ukb_code_mappings,
  input_icd10_format = "ICD10_CODE",
  output_icd10_format = "ALT_CODE"
)

Arguments

icd10_codes
character vector of ICD-10 codes

ukb_code_mappings
the ukb_code_mappings list returned by get_ukb_code_mappings.

input_icd10_format
character. Must be either "ICD10_CODE" or "ALT_CODE".

output_icd10_format
character. Must be either "ICD10_CODE" or "ALT_CODE".

Value

character vector of ICD-10 codes, reformatted as specified by output_icd10_format.

See also