Reformat a dataframe of clinical codes to work with extract_first_or_last_clinical_event_multi

A utility function that helps reformat the output from map_codes or lookup_codes to work with extract_first_or_last_clinical_event_multi. See also output from generate_self_reported_diabetes_codes_df for an example of the format that this function will output.

reformat_standardised_codelist(
  standardised_codelist,
  code_type,
  disease,
  disease_category,
  phenotype_source
)

Arguments

standardised_codelist
a data frame with column names "code", "description", "code_type".

code_type
character (scalar). The clinical code type e.g. "read2"

disease
character (scalar), e.g. "Secondary polycythaemia"

disease_category
character (scalar). The subcategory of disease that these codes belong to e.g. "Diagnosis of Secondary polycythaemia".

phenotype_source
character (scalar), e.g. "caliber".

Value

A data frame with the following column names: 'disease', 'description', 'category', 'code_type', 'code' and 'phenotype_source'.