Summarise a UK Biobank phenotype dataset row-wise

Allows the user to summarise row-wise across a set of column "groups". e.g. calculate standard deviation values for all systolic and diastolic blood pressure readings per eid. Can also accept custom functions (e.g. a function to count the number of non-NA values). Custom functions should include a '...' parameter.

summarise_rowise(
  ukb_pheno,
  functions,
  data_dict,
  grouping_col = "Field_FieldID",
  selected_col_groups,
  ...
)

Arguments

ukb_pheno
a UK Biobank phenotype dataset

functions
a character vector of function names (e.g. c("sd"))

data_dict
a data dictionary for ukb_pheno

grouping_col
the name of a column in data_dict that indicates column groups to summarise

selected_col_groups
a character vector of column groups to summarise (i.e. can be a subset of the unique values in data_dict$grouping_col)

...
arguments to passed on to summary functions listed in functions

Examples

if (FALSE) {

# make a custom function that counts the number of non-NA values.
# Note: include '...', otherwise an error is thrown by including na.rm = TRUE below

n_not_na <- function(x, ...) {
sum(!is.na(x))
}

# summarise all numerical columns in UKB dataset by calculating mean/sd/n_not_na,
# including argument na.rm = TRUE

summarise_test_read_pheno <- summarise_rowise(
ukb_pheno = ukb_pheno, # UKB phenotype dataset as a datatable
functions = c("mean", "sd", "n_not_na"),
data_dict = data_dict, # data dictionary generated by make_data_dict()
grouping_col = "Field_FieldID", # to summarise by FieldID
selected_col_groups = data_dict %>% # ...and only summarise numerical data
  dplyr::filter(ValueType %in% c("Continuous", "Integer")) %>%
  .$Field_FieldID %>%
  unique(),
na.rm = TRUE
)
}