Derive simplified ethnic background

Simplifies ethnic background in a UK Biobank main dataset to the main categories for Field ID 21000.

derive_ethnic_background_simplified(
  ukb_main,
  ukb_data_dict = get_ukb_data_dict(),
  ethnicity_levels = c("White", "Mixed", "Asian or Asian British",
    "Black or Black British", "Chinese", "Other ethnic group"),
  .drop = FALSE,
  .details_only = FALSE
)

Arguments

ukb_main
A UK Biobank main dataset.

ukb_data_dict
The UKB data dictionary (available online at the UK Biobank data showcase. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

ethnicity_levels
The factor level order for the appended ethnic_background_simplified column. By default, the baseline level is set to "White" ethnicity.

.drop
If TRUE, remove the required input columns from the result

.details_only
If TRUE, return a list containing details of required input variables (Field IDs) and derived variables (new column name, label and values/value labels).

Value

A data frame with a column called ethnic_background_simplified (type factor).

Details

Categories "Do not know" and "Prefer not to answer" are converted to NA. A new column called ethnic_background_simplified of type factor is appended to the input data frame. By default, "White" ethnicity is set to the baseline level as this is the largest category. Levels can be explicitly specified using the ethnicity_levels argument.

Examples

dummy_ukb_data <- data.frame(
  eid = c(1, 2, 3),
  f.21000.0.0 = factor(c("White", "White and Black Caribbean", NA)),
  f.21000.1.0 = factor(c("British", "Caribbean", NA)),
  f.21000.2.0 = factor(c("Irish", NA, "Any other Asian background"))
)

derive_ethnic_background_simplified(
  ukb_main = dummy_ukb_data,
  ukb_data_dict = get_ukb_data_dict()
)
#>   eid               f.21000.0.0 f.21000.1.0                f.21000.2.0
#> 1   1                     White     British                      Irish
#> 2   2 White and Black Caribbean   Caribbean                       <NA>
#> 3   3                      <NA>        <NA> Any other Asian background
#>   ethnic_background_simplified
#> 1                        White
#> 2                        Mixed
#> 3       Asian or Asian British