Extract phenotypes from clinical events data

Filters a clinical events table created by tidy_clinical_events for a set or sets of specified clinical codes that represent one or more phenotypes. By default, the earliest date that any clinical code appears in an individual participant's record is extracted. See also the 'Clinical events' vignette on the ukbwranglr package website.

extract_phenotypes(
  clinical_events,
  clinical_codes,
  data_sources = NULL,
  min_max = "min",
  colnames_prefix = NULL,
  labels_prefix = NULL,
  workers = NULL,
  keep_all = FALSE
)

Arguments

clinical_events
A long format data frame created by tidy_clinical_events, tidy_gp_clinical, tidy_gp_scripts or make_clinical_events_db. This can also be a tbl_dbi object.

clinical_codes
data frame. Must match the format as per example_clinical_codes.

data_sources
A character vector of clinical events sources in clinical_events to extract phenotypes from. Use clinical_events_sources (source column) for a list of valid values.

min_max
Choose either "min" or "max" to extract either the earliest or latest date that one of the codes appears in clinical_codes appears.

colnames_prefix
Optionally add a prefix to column names.

labels_prefix
Optionally add a prefix to variable labels.

workers
The number of processes to run in parallel when clinical_codes includes multiple diseases. This is used to set the number of workers in plan with strategy = multisession, which is passed on to future_map. If NULL, then no processing is performed sequentially. Default value is NULL.

keep_all
If TRUE, keep all matching codes for each eid in a nested data frame column named 'data'.

Value

A named list of data frames, one for each disease. Each data frame has an "eid" column, and "event_min/max_indicator" and "event_min/max_date" columns for each phenotype in the 'category' column of clinical_codes for that disease. If keep_all is TRUE, then there will also be additional nested data frame column called 'data'.

Examples

if (FALSE) {
dummy_main_dataset_clinical_events() %>%
  tidy_clinical_events(clinical_events_sources = c("summary_hes_icd10")) %>%
  dplyr::bind_rows() %>%
  extract_phenotypes(
    clinical_codes = example_clinical_codes(),
    data_sources = "f41270",
    min_max = "min"
  )
}