Label a UK Biobank main dataset

Adds variable and value labels a UK Biobank main dataset using labelled. Note that labels are only supported for numeric (integer or double) and character vectors.

label_ukb_main(
  ukb_main,
  data_dict,
  ukb_codings = get_ukb_codings(),
  colnames_col = "descriptive_colnames",
  max_n_labels = 30
)

Arguments

ukb_main

A UK Biobank main dataset (data frame)

data_dict

A data dictionary specific to the UKB main dataset file, generated by make_data_dict. This can be filtered for a subset of columns to be labelled.

ukb_codings

The UKB codings dictionary. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

colnames_col

The name of the column in data_dict that contains column names matching those in ukb_main.

max_n_labels

Coded variables with associated value labels up to (and equal to) this threshold will be labelled using. If NULL, then all value labels will be applied. Default value is 30. labelled if labelled is TRUE.

Value

A data frame with labelled

Examples

if (FALSE) {
dummy_data <- dummy_main_dataset_clinical_events()

label_ukb_main(
  ukb_main = dummy_data,
  data_dict = make_data_dict(dummy_data)
)
}