Create a SQLite database with a clinical_events table

Adds tables called clinical_events and gp_clinical_values to a SQLite database file. This is a long format table combining all clinical events data from a UK Biobank main dataset and the UK Biobank primary care clinical events dataset, as listed by link{clinical_events_sources}. Indexes are set on the source, code and eid columns in the clinical_events table.

make_clinical_events_db(
  ukb_main_path,
  gp_clinical_path,
  gp_scripts_path,
  ukb_db_path,
  ukb_main_delim = "\t",
  ukb_data_dict = get_ukb_data_dict(),
  ukb_codings = get_ukb_codings(),
  overwrite = FALSE,
  strict = TRUE,
  chunk_size = 10000
)

Arguments

ukb_main_path

Path to the main UKB dataset file.

gp_clinical_path

Path to the UKB primary care clinical events file (gp_clinical.txt).

gp_scripts_path

Path to the UKB primary care prescriptions file (gp_scripts.txt).

ukb_db_path

Path to the SQLite database file. The file name must end with '.db'. If no file with this name exists then one will be created.

ukb_main_delim

Delimiter for ukb_main_path. Default value is "\t".

ukb_data_dict

The UKB data dictionary (available online at the UK Biobank data showcase. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

ukb_codings

The UKB codings dictionary. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

overwrite

If TRUE, then tables clinical_events and gp_clinical_values will be overwritten if they already exist in the database. Default value is FALSE.

strict

If FALSE, create database regardless of whether the main UKB dataset file is missing any required clinical events fields (see tidy_clinical_events for further details). Default is TRUE.

chunk_size

The number of rows to include in each chunk when processing the primary care dataset.

Value

Returns NULL invisibly

See also