Summarise numerical variables

Summarises numerical variables with repeated measurements either by field (i.e. all available measurements) or by instance (i.e. for all measurements at each assessment visit). Currently available summary options are mean, minimum, maximum, sum and number of non-missing values.

summarise_numerical_variables(
  ukb_main,
  data_dict = NULL,
  summary_function = "mean",
  summarise_by = "Field",
  .drop = FALSE
)

Arguments

ukb_main

A UK Biobank main dataset data frame. Column names must match those under the descriptive_colnames column in data_dict.

data_dict

a data dictionary specific to the UKB main dataset file, created by make_data_dict.

summary_function

The summary function to be applied. Options: "mean", "min", "max", "sum" or "n_values"

summarise_by

Whether to summarise by "Field" or by "Instance".

.drop

If TRUE, removes the original numerical variables from the result. Default value is FALSE.

Value

A data frame with new columns summarising numerical variables. The names for these new columns are prefixed by the value for summary_function and end with 'x', FieldID +/- instance being summarised e.g. if summarising FieldID 4080 instance 0, the new column would be named 'mean_systolic_blood_pressure_automated_reading_x4080_0'.

Details

Note that when summary_function = "sum", missing values are converted to zero. Therefore if a set of values are all missing then the sum will summarised as 0. See the documentation for rowSums for further details.

Examples

# make dummy data
dummy_ukb_data <- tibble::tibble(
  f.eid = c(1, 2, 3, 4, 5, 6),
  f.31.0.0 = c(0, 0, 1, 0, NA, 1),
  f.34.0.0 = c(1952, 1946, 1951, 1956, NA, 1948),
  f.52.0.0 = c(8, 3, 4, 9, 4, 2),
  f.21000.0.0 = c(NA, 4001, 3, NA, -3, 3),
  f.20002.0.0 = c(1665, 1383, 1197, 1441, 1429, 1513),
  f.21001.0.0 = c(20.1115, 30.1536, 22.8495, 23.4904, 29.2752, 28.2567),
  f.21001.1.0 = c(20.864, 20.2309, 26.7929, 25.6826, 19.7576, 30.286),
  f.21001.2.0 = c(NA, 27.4936, 27.6286, 37.2294, 14.6641, 27.3534),
)

# summarise, keep original variables
summarise_numerical_variables(dummy_ukb_data)
#> Number of summary columns to make: 1
#> Time taken: 0 minutes, 0 seconds.
#> # A tibble: 6 × 10
#>   f.eid f.31.0.0 f.34.0.0 f.52.0.0 f.21000.0.0 f.20002.0.0 f.21001.0.0
#>   <dbl>    <dbl>    <dbl>    <dbl>       <dbl>       <dbl>       <dbl>
#> 1     1        0     1952        8          NA        1665        20.1
#> 2     2        0     1946        3        4001        1383        30.2
#> 3     3        1     1951        4           3        1197        22.8
#> 4     4        0     1956        9          NA        1441        23.5
#> 5     5       NA       NA        4          -3        1429        29.3
#> 6     6        1     1948        2           3        1513        28.3
#> # … with 3 more variables: f.21001.1.0 <dbl>, f.21001.2.0 <dbl>,
#> #   mean_body_mass_index_bmi_x21001 <dbl+lbl>

# summarise, drop original variables
summarise_numerical_variables(dummy_ukb_data, .drop = TRUE)
#> Number of summary columns to make: 1
#> Time taken: 0 minutes, 0 seconds.
#> # A tibble: 6 × 7
#>   f.eid f.31.0.0 f.34.0.0 f.52.0.0 f.21000.0.0 f.20002.0.0 mean_body_mass_index…
#>   <dbl>    <dbl>    <dbl>    <dbl>       <dbl>       <dbl>             <dbl+lbl>
#> 1     1        0     1952        8          NA        1665                  20.5
#> 2     2        0     1946        3        4001        1383                  26.0
#> 3     3        1     1951        4           3        1197                  25.8
#> 4     4        0     1956        9          NA        1441                  28.8
#> 5     5       NA       NA        4          -3        1429                  21.2
#> 6     6        1     1948        2           3        1513                  28.6