Tidy clinical events data from a UK Biobank main dataset

Data in a UK Biobank main dataset is stored in wide format i.e. a single row of data per UK Biobank participant ('eid's). Clinical events may be ascertained from numerous sources (e.g. self-reported medical conditions, linked hospital records) with coded events and their associated dates recorded across multiple columns. This function tidies this data into a standardised long format table.

tidy_clinical_events(
  ukb_main,
  ukb_data_dict = get_ukb_data_dict(),
  ukb_codings = get_ukb_codings(),
  clinical_events_sources = c("primary_death_icd10", "secondary_death_icd10",
    "self_report_medication", "self_report_non_cancer", "self_report_non_cancer_icd10",
    "self_report_cancer", "self_report_operation", "cancer_register_icd9",
    "cancer_register_icd10", "summary_hes_icd9", "summary_hes_icd10",
    "summary_hes_opcs3", "summary_hes_opcs4"),
  strict = TRUE,
  .details_only = FALSE
)

Arguments

ukb_main

A UK Biobank main dataset.

ukb_data_dict

The UKB data dictionary (available online at the UK Biobank data showcase. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

ukb_codings

The UKB codings dictionary. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

clinical_events_sources

A character vector of clinical events sources to tidy. By default, all available options are included.

strict

If TRUE, raise an error if required columns for any clinical events sources listed in clinical_events are not present in ukb_main. If FALSE, then a warning message will be displayed instead. Default value is TRUE.

.details_only

If TRUE, return a list detailing required Field IDs

Value

A named list of clinical events data frames.

Details

A named list of data frames is returned, with the names corresponding to the data sources specified by clinical_events. Each data frame has the following columns:

eid - participant identifier

source - the FieldID (prefixed by 'f') where clinical codes were extracted from. See clinical_events_sources for further details.

index - the corresponding instance and array (e.g. '0-1' means instance 0 and array 1)

code - clinical code. The type of clinical codings system used depends on source.

date - associated date. Note that in cases where participants self-reported a medical condition but recorded the date as either 'Date uncertain or unknown' or 'Preferred not to answer' (see data coding 13) then the date is set to NA.

Other notes

Results may be combined into a single data frame using bind_rows.

Examples

# dummy UKB main dataset and metadata
dummy_ukb_main <- get_ukb_dummy("dummy_ukb_main.tsv")
dummy_ukb_data_dict <- get_ukb_dummy("dummy_Data_Dictionary_Showcase.tsv")
dummy_ukb_codings <- get_ukb_dummy("dummy_Codings.tsv")

# tidy clinical events in a UK Biobank main dataset
tidy_clinical_events(
  ukb_main = dummy_ukb_main,
  ukb_data_dict = dummy_ukb_data_dict,
  ukb_codings = dummy_ukb_codings
)
#> Tidying clinical events for primary_death_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for secondary_death_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_medication
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_non_cancer
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_non_cancer_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_cancer
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_operation
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for cancer_register_icd9
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for cancer_register_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_icd9
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_opcs3
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_opcs4
#> Time taken: 0 minutes, 0 seconds.
#> $primary_death_icd10
#>    eid source index code       date
#> 1:   1 f40001    NA X095 1917-10-08
#> 2:   2 f40001    NA A162 1955-02-11
#> 3:   1 f40001    NA X095 1910-02-19
#> 4:   2 f40001    NA A162 1965-08-08
#> 5:   1 f40001    NA X095 1917-10-08
#> 6:   2 f40001    NA A162 1955-02-11
#> 7:   1 f40001    NA X095 1910-02-19
#> 8:   2 f40001    NA A162 1965-08-08
#> 
#> $secondary_death_icd10
#>    eid source index code       date
#> 1:   1 f40002    NA W192 1917-10-08
#> 2:   2 f40002    NA V374 1955-02-11
#> 3:   1 f40002    NA W192 1910-02-19
#> 4:   2 f40002    NA V374 1965-08-08
#> 5:   1 f40002    NA X715 1917-10-08
#> 6:   1 f40002    NA X715 1910-02-19
#> 
#> $self_report_medication
#>     eid source index       code       date
#>  1:   1 f20003    NA 1140861958 1955-02-11
#>  2:   2 f20003    NA 1141146234 1965-08-08
#>  3:   1 f20003    NA 1140861958 1910-02-19
#>  4:   2 f20003    NA 1141146234 1915-03-18
#>  5:   1 f20003    NA 1141146188 1955-02-11
#>  6:   2 f20003    NA 1141184722 1965-08-08
#>  7:   1 f20003    NA 1141146188 1910-02-19
#>  8:   2 f20003    NA 1141184722 1915-03-18
#>  9:   1 f20003    NA 1141184722 1955-02-11
#> 10:   2 f20003    NA 1140861958 1965-08-08
#> 11:   1 f20003    NA 1141184722 1910-02-19
#> 12:   2 f20003    NA 1140861958 1915-03-18
#> 
#> $self_report_non_cancer
#>     eid source index code       date
#>  1:   1 f20002    NA 1665 1998-12-24
#>  2:   2 f20002    NA 1383 2011-01-05
#>  3:   3 f20002    NA 1665       <NA>
#>  4:   4 f20002    NA 1383       <NA>
#>  5:   1 f20002    NA 1665 2003-02-25
#>  6:   2 f20002    NA 1383 2020-07-02
#>  7:   3 f20002    NA 1665       <NA>
#>  8:   4 f20002    NA 1383       <NA>
#>  9:   1 f20002    NA 1665 2011-04-07
#> 10:   2 f20002    NA 1383 1981-03-01
#> 11:   3 f20002    NA 1665       <NA>
#> 12:   4 f20002    NA 1383       <NA>
#> 13:   1 f20002    NA 1665 2018-10-14
#> 14:   2 f20002    NA 1383 1983-01-03
#> 15:   3 f20002    NA 1665       <NA>
#> 16:   4 f20002    NA 1383       <NA>
#> 17:   1 f20002    NA 1223 1998-12-24
#> 18:   2 f20002    NA 1352 2011-01-05
#> 19:   1 f20002    NA 1223 2003-02-25
#> 20:   2 f20002    NA 1352 2020-07-02
#> 21:   1 f20002    NA 1223 2011-04-07
#> 22:   2 f20002    NA 1352 1981-03-01
#> 23:   1 f20002    NA 1223 2018-10-14
#> 24:   2 f20002    NA 1352 1983-01-03
#> 25:   1 f20002    NA 1514 1998-12-24
#> 26:   2 f20002    NA 1447 2011-01-05
#> 27:   1 f20002    NA 1514 2003-02-25
#> 28:   2 f20002    NA 1447 2020-07-02
#> 29:   1 f20002    NA 1514 2011-04-07
#> 30:   2 f20002    NA 1447 1981-03-01
#> 31:   1 f20002    NA 1514 2018-10-14
#> 32:   2 f20002    NA 1447 1983-01-03
#> 33:   2 f20002    NA 1165 2011-01-05
#> 34:   2 f20002    NA 1165 2020-07-02
#> 35:   2 f20002    NA 1165 1981-03-01
#> 36:   2 f20002    NA 1165 1983-01-03
#>     eid source index code       date
#> 
#> $self_report_non_cancer_icd10
#>     eid       source index code       date
#>  1:   1 f20002_icd10    NA  N95 1998-12-24
#>  2:   2 f20002_icd10    NA  M33 2011-01-05
#>  3:   3 f20002_icd10    NA  N95       <NA>
#>  4:   4 f20002_icd10    NA  M33       <NA>
#>  5:   1 f20002_icd10    NA  N95 2003-02-25
#>  6:   2 f20002_icd10    NA  M33 2020-07-02
#>  7:   3 f20002_icd10    NA  N95       <NA>
#>  8:   4 f20002_icd10    NA  M33       <NA>
#>  9:   1 f20002_icd10    NA  N95 2011-04-07
#> 10:   2 f20002_icd10    NA  M33 1981-03-01
#> 11:   3 f20002_icd10    NA  N95       <NA>
#> 12:   4 f20002_icd10    NA  M33       <NA>
#> 13:   1 f20002_icd10    NA  N95 2018-10-14
#> 14:   2 f20002_icd10    NA  M33 1983-01-03
#> 15:   3 f20002_icd10    NA  N95       <NA>
#> 16:   4 f20002_icd10    NA  M33       <NA>
#> 17:   1 f20002_icd10    NA  E11 1998-12-24
#> 18:   2 f20002_icd10    NA  N84 2011-01-05
#> 19:   1 f20002_icd10    NA  E11 2003-02-25
#> 20:   2 f20002_icd10    NA  N84 2020-07-02
#> 21:   1 f20002_icd10    NA  E11 2011-04-07
#> 22:   2 f20002_icd10    NA  N84 1981-03-01
#> 23:   1 f20002_icd10    NA  E11 2018-10-14
#> 24:   2 f20002_icd10    NA  N84 1983-01-03
#> 25:   1 f20002_icd10    NA  N30 1998-12-24
#> 26:   2 f20002_icd10    NA  D61 2011-01-05
#> 27:   1 f20002_icd10    NA  N30 2003-02-25
#> 28:   2 f20002_icd10    NA  D61 2020-07-02
#> 29:   1 f20002_icd10    NA  N30 2011-04-07
#> 30:   2 f20002_icd10    NA  D61 1981-03-01
#> 31:   1 f20002_icd10    NA  N30 2018-10-14
#> 32:   2 f20002_icd10    NA  D61 1983-01-03
#> 33:   2 f20002_icd10    NA  K85 2011-01-05
#> 34:   2 f20002_icd10    NA  K85 2020-07-02
#> 35:   2 f20002_icd10    NA  K85 1981-03-01
#> 36:   2 f20002_icd10    NA  K85 1983-01-03
#>     eid       source index code       date
#> 
#> $self_report_cancer
#>     eid source index code       date
#>  1:   1 f20001    NA 1048 2012-10-26
#>  2:   2 f20001    NA 1046 2016-01-24
#>  3:   1 f20001    NA 1048 2007-02-01
#>  4:   2 f20001    NA 1046 2023-03-01
#>  5:   1 f20001    NA 1048 2023-03-16
#>  6:   2 f20001    NA 1046 2024-01-14
#>  7:   1 f20001    NA 1048 2014-09-27
#>  8:   2 f20001    NA 1046 2013-03-16
#>  9:   1 f20001    NA 1005 2012-10-26
#> 10:   2 f20001    NA 1003 2016-01-24
#> 11:   1 f20001    NA 1005 2007-02-01
#> 12:   2 f20001    NA 1003 2023-03-01
#> 13:   1 f20001    NA 1005 2023-03-16
#> 14:   2 f20001    NA 1003 2024-01-14
#> 15:   1 f20001    NA 1005 2014-09-27
#> 16:   2 f20001    NA 1003 2013-03-16
#> 17:   1 f20001    NA 1045 2012-10-26
#> 18:   2 f20001    NA 1028 2016-01-24
#> 19:   1 f20001    NA 1045 2007-02-01
#> 20:   2 f20001    NA 1028 2023-03-01
#> 21:   1 f20001    NA 1045 2023-03-16
#> 22:   2 f20001    NA 1028 2024-01-14
#> 23:   1 f20001    NA 1045 2014-09-27
#> 24:   2 f20001    NA 1028 2013-03-16
#> 25:   1 f20001    NA 1017 2012-10-26
#> 26:   2 f20001    NA 1039 2016-01-24
#> 27:   1 f20001    NA 1017 2007-02-01
#> 28:   2 f20001    NA 1039 2023-03-01
#> 29:   1 f20001    NA 1017 2023-03-16
#> 30:   2 f20001    NA 1039 2024-01-14
#> 31:   1 f20001    NA 1017 2014-09-27
#> 32:   2 f20001    NA 1039 2013-03-16
#>     eid source index code       date
#> 
#> $self_report_operation
#>    eid source index code       date
#> 1:   1 f20004    NA 1102 2012-10-26
#> 2:   2 f20004    NA 1105 2016-01-24
#> 3:   1 f20004    NA 1102 2008-03-26
#> 4:   2 f20004    NA 1105       <NA>
#> 5:   1 f20004    NA 1108 2012-10-26
#> 6:   2 f20004    NA 1109 2016-01-24
#> 7:   1 f20004    NA 1108 2008-03-26
#> 8:   2 f20004    NA 1109       <NA>
#> 
#> $cancer_register_icd9
#>    eid source index   code       date
#> 1:   1 f40013    NA  27134 1956-11-24
#> 2:   2 f40013    NA   9626 1910-10-04
#> 3:   1 f40013    NA  27134 1962-09-04
#> 4:   2 f40013    NA   9626       <NA>
#> 5:   1 f40013    NA   2042 1956-11-24
#> 6:   2 f40013    NA E90200 1910-10-04
#> 7:   1 f40013    NA   2042 1962-09-04
#> 8:   2 f40013    NA E90200       <NA>
#> 
#> $cancer_register_icd10
#>    eid source index  code       date
#> 1:   1 f40006    NA M4815 1956-11-24
#> 2:   1 f40006    NA M4815 1962-09-04
#> 3:   1 f40006    NA  C850 1956-11-24
#> 4:   2 f40006    NA  W192 1910-10-04
#> 5:   1 f40006    NA  C850 1962-09-04
#> 6:   2 f40006    NA  W192       <NA>
#> 
#> $summary_hes_icd9
#>    eid source index   code       date
#> 1:   1 f41271    NA E89115 1917-10-08
#> 2:   2 f41271    NA  E8326 1955-02-11
#> 3:   1 f41271    NA E89115 1969-11-23
#> 4:   2 f41271    NA  E8326 1956-09-12
#> 5:   2 f41271    NA  75513 1955-02-11
#> 6:   2 f41271    NA  75513 1956-09-12
#> 
#> $summary_hes_icd10
#>    eid source index  code       date
#> 1:   1 f41270    NA  X715 1955-11-12
#> 2:   2 f41270    NA   E11 1939-02-16
#> 3:   1 f41270    NA  X715 1910-02-19
#> 4:   2 f41270    NA   E11 1965-08-08
#> 5:   1 f41270    NA   E10 1955-11-12
#> 6:   2 f41270    NA M0087 1939-02-16
#> 7:   1 f41270    NA   E10 1910-02-19
#> 8:   2 f41270    NA M0087 1965-08-08
#> 
#> $summary_hes_opcs3
#>    eid source index code       date
#> 1:   1 f41273    NA  001 1969-11-23
#> 2:   2 f41273    NA 0011 1956-09-12
#> 3:   1 f41273    NA  001 1955-11-12
#> 4:   2 f41273    NA 0011 1939-02-16
#> 5:   1 f41273    NA 0081 1969-11-23
#> 6:   2 f41273    NA 0071 1956-09-12
#> 7:   1 f41273    NA 0081 1955-11-12
#> 8:   2 f41273    NA 0071 1939-02-16
#> 
#> $summary_hes_opcs4
#>     eid source index code       date
#>  1:   1 f41272    NA  A01 1956-11-24
#>  2:   2 f41272    NA A023 1910-10-04
#>  3:   3 f41272    NA  H01       <NA>
#>  4:   4 f41272    NA H011       <NA>
#>  5:   5 f41272    NA H022       <NA>
#>  6:   6 f41272    NA H013       <NA>
#>  7:   7 f41272    NA H018       <NA>
#>  8:   8 f41272    NA H019       <NA>
#>  9:   1 f41272    NA  A01 1969-11-23
#> 10:   2 f41272    NA A023 1956-09-12
#> 11:   3 f41272    NA  H01       <NA>
#> 12:   4 f41272    NA H011       <NA>
#> 13:   5 f41272    NA H022       <NA>
#> 14:   6 f41272    NA H013       <NA>
#> 15:   7 f41272    NA H018       <NA>
#> 16:   8 f41272    NA H019       <NA>
#> 17:   1 f41272    NA A018 1956-11-24
#> 18:   2 f41272    NA  A02 1910-10-04
#> 19:   1 f41272    NA A018 1969-11-23
#> 20:   2 f41272    NA  A02 1956-09-12
#> 

# use .details_only = TRUE to return details of required Field IDs for
# specific clinical_events sources
tidy_clinical_events(.details_only = TRUE)
#> $required_field_ids
#> $required_field_ids$primary_death_icd10
#> code_fid date_fid 
#>  "40001"  "40000" 
#> 
#> $required_field_ids$secondary_death_icd10
#> code_fid date_fid 
#>  "40002"  "40000" 
#> 
#> $required_field_ids$self_report_medication
#> code_fid date_fid 
#>  "20003"     "53" 
#> 
#> $required_field_ids$self_report_non_cancer
#> code_fid date_fid 
#>  "20002"  "20008" 
#> 
#> $required_field_ids$self_report_non_cancer_icd10
#> code_fid date_fid 
#>  "20002"  "20008" 
#> 
#> $required_field_ids$self_report_cancer
#> code_fid date_fid 
#>  "20001"  "20006" 
#> 
#> $required_field_ids$self_report_operation
#> code_fid date_fid 
#>  "20004"  "20010" 
#> 
#> $required_field_ids$cancer_register_icd9
#> code_fid date_fid 
#>  "40013"  "40005" 
#> 
#> $required_field_ids$cancer_register_icd10
#> code_fid date_fid 
#>  "40006"  "40005" 
#> 
#> $required_field_ids$summary_hes_icd9
#> code_fid date_fid 
#>  "41271"  "41281" 
#> 
#> $required_field_ids$summary_hes_icd10
#> code_fid date_fid 
#>  "41270"  "41280" 
#> 
#> $required_field_ids$summary_hes_opcs3
#> code_fid date_fid 
#>  "41273"  "41283" 
#> 
#> $required_field_ids$summary_hes_opcs4
#> code_fid date_fid 
#>  "41272"  "41282" 
#> 
#>