Tidy clinical events data from a UK Biobank main dataset

Data in a UK Biobank main dataset is stored in wide format i.e. a single row of data per UK Biobank participant ('eid's). Clinical events may be ascertained from numerous sources (e.g. self-reported medical conditions, linked hospital records) with coded events and their associated dates recorded across multiple columns. This function tidies this data into a standardised long format table.

tidy_clinical_events(
  ukb_main,
  ukb_data_dict = get_ukb_data_dict(),
  ukb_codings = get_ukb_codings(),
  clinical_events_sources = c("primary_death_icd10", "secondary_death_icd10",
    "self_report_medication", "self_report_non_cancer", "self_report_non_cancer_icd10",
    "self_report_cancer", "self_report_operation", "cancer_register_icd9",
    "cancer_register_icd10", "summary_hes_icd9", "summary_hes_icd10",
    "summary_hes_opcs3", "summary_hes_opcs4"),
  strict = TRUE,
  .details_only = FALSE
)

Arguments

ukb_main
A UK Biobank main dataset. Must be a data.table

ukb_data_dict
The UKB data dictionary (available online at the UK Biobank data showcase. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

ukb_codings
The UKB codings dictionary. This should be a data frame where all columns are of type character. By default, this is downloaded from the ukbwranglr_resources github repo.

clinical_events_sources
A character vector of clinical events sources to tidy. By default, all available options are included.

strict
If TRUE, raise an error if required columns for any clinical events sources listed in clinical_events are not present in ukb_main. If FALSE, then a warning message will be displayed instead. Default value is TRUE.

.details_only
If TRUE, return a list detailing required Field IDs

Value

A named list of clinical events data frames.

Details

A named list of data frames is returned, with the names corresponding to the data sources specified by clinical_events. Each data frame has the following columns:

eid - participant identifier

source - the FieldID (prefixed by 'f') where clinical codes were extracted from. See clinical_events_sources for further details.

index - the corresponding instance and array (e.g. '0-1' means instance 0 and array 1)

code - clinical code. The type of clinical codings system used depends on source.

date - associated date. Note that in cases where participants self-reported a medical condition but recorded the date as either 'Date uncertain or unknown' or 'Preferred not to answer' (see data coding 13) then the date is set to NA.

Other notes

Results may be combined into a single data frame using bind_rows.

Examples

# tidy clinical events in a UK Biobank main dataset
tidy_clinical_events(
  ukb_main = dummy_main_dataset_clinical_events(),
  ukb_data_dict = get_ukb_data_dict(),
  ukb_codings = get_ukb_codings()
)
#> Tidying clinical events for primary_death_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for secondary_death_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_medication
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_non_cancer
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_non_cancer_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_cancer
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for self_report_operation
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for cancer_register_icd9
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for cancer_register_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_icd9
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_icd10
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_opcs3
#> Time taken: 0 minutes, 0 seconds.
#> Tidying clinical events for summary_hes_opcs4
#> Time taken: 0 minutes, 0 seconds.
#> $primary_death_icd10
#>    eid source index code       date
#> 1:   1 f40001   0_0 X095 1917-10-08
#> 2:   2 f40001   0_0 A162 1955-02-11
#> 3:   1 f40001   1_0 X095 1910-02-19
#> 4:   2 f40001   1_0 A162 1965-08-08
#> 
#> $secondary_death_icd10
#>    eid source index code       date
#> 1:   1 f40002   0_0 W192 1917-10-08
#> 2:   2 f40002   0_0 V374 1955-02-11
#> 3:   1 f40002   1_3 X715 1910-02-19
#> 
#> $self_report_medication
#>    eid source index       code       date
#> 1:   1 f20003   0_0 1140861958 1955-02-11
#> 2:   2 f20003   0_0 1141146234 1965-08-08
#> 3:   1 f20003   2_0 1141146188 1910-02-19
#> 4:   2 f20003   2_0 1141184722 1915-03-18
#> 5:   1 f20003   2_3 1141184722 1910-02-19
#> 6:   2 f20003   2_3 1140861958 1915-03-18
#> 
#> $self_report_non_cancer
#>    eid source index code       date
#> 1:   1 f20002   0_0 1665 1998-12-24
#> 2:   2 f20002   0_0 1383 2011-01-05
#> 3:   1 f20002   0_3 1223 2003-02-25
#> 4:   2 f20002   0_3 1352 2020-07-02
#> 5:   1 f20002   2_0 1514 2011-04-07
#> 6:   2 f20002   2_0 1447 1981-03-01
#> 7:   2 f20002   2_3 1165 1983-01-03
#> 
#> $self_report_non_cancer_icd10
#>    eid       source index code       date
#> 1:   1 f20002_icd10   0_0  N95 1998-12-24
#> 2:   2 f20002_icd10   0_0  M33 2011-01-05
#> 3:   1 f20002_icd10   0_3  E11 2003-02-25
#> 4:   2 f20002_icd10   0_3  N84 2020-07-02
#> 5:   1 f20002_icd10   2_0  N30 2011-04-07
#> 6:   2 f20002_icd10   2_0  D61 1981-03-01
#> 7:   2 f20002_icd10   2_3  K85 1983-01-03
#> 
#> $self_report_cancer
#>    eid source index code       date
#> 1:   1 f20001   0_0 1048 2012-10-26
#> 2:   2 f20001   0_0 1046 2016-01-24
#> 3:   1 f20001   0_3 1005 2007-02-01
#> 4:   2 f20001   0_3 1003 2023-03-01
#> 5:   1 f20001   2_0 1045 2023-03-16
#> 6:   2 f20001   2_0 1028 2024-01-14
#> 7:   1 f20001   2_3 1017 2014-09-27
#> 8:   2 f20001   2_3 1039 2013-03-16
#> 
#> $self_report_operation
#>    eid source index code       date
#> 1:   1 f20004   0_0 1102 2012-10-26
#> 2:   2 f20004   0_0 1105 2016-01-24
#> 3:   1 f20004   0_3 1108 2008-03-26
#> 4:   2 f20004   0_3 1109       <NA>
#> 
#> $cancer_register_icd9
#>    eid source index   code       date
#> 1:   1 f40013   0_0  27134 1956-11-24
#> 2:   2 f40013   0_0   9626 1910-10-04
#> 3:   1 f40013   2_0   2042 1962-09-04
#> 4:   2 f40013   2_0 E90200       <NA>
#> 
#> $cancer_register_icd10
#>    eid source index  code       date
#> 1:   1 f40006   0_0 M4815 1956-11-24
#> 2:   1 f40006   2_0  C850 1962-09-04
#> 3:   2 f40006   2_0  W192       <NA>
#> 
#> $summary_hes_icd9
#>    eid source index   code       date
#> 1:   1 f41271   0_0 E89115 1917-10-08
#> 2:   2 f41271   0_0  E8326 1955-02-11
#> 3:   2 f41271   0_3  75513 1956-09-12
#> 
#> $summary_hes_icd10
#>    eid source index  code       date
#> 1:   1 f41270   0_0  X715 1955-11-12
#> 2:   2 f41270   0_0   E11 1939-02-16
#> 3:   1 f41270   0_3   E10 1910-02-19
#> 4:   2 f41270   0_3 M0087 1965-08-08
#> 
#> $summary_hes_opcs3
#>    eid source index code       date
#> 1:   1 f41273   0_0  001 1969-11-23
#> 2:   2 f41273   0_0 0011 1956-09-12
#> 3:   1 f41273   0_3 0081 1955-11-12
#> 4:   2 f41273   0_3 0071 1939-02-16
#> 
#> $summary_hes_opcs4
#>    eid source index code       date
#> 1:   1 f41272   0_0  A01 1956-11-24
#> 2:   2 f41272   0_0 A023 1910-10-04
#> 3:   1 f41272   0_3 A018 1969-11-23
#> 4:   2 f41272   0_3  A02 1956-09-12
#> 

# use .details_only = TRUE to return details of required Field IDs for
# specific clinical_events sources
tidy_clinical_events(.details_only = TRUE)
#> $required_field_ids
#> $required_field_ids$primary_death_icd10
#> code_fid date_fid 
#>  "40001"  "40000" 
#> 
#> $required_field_ids$secondary_death_icd10
#> code_fid date_fid 
#>  "40002"  "40000" 
#> 
#> $required_field_ids$self_report_medication
#> code_fid date_fid 
#>  "20003"     "53" 
#> 
#> $required_field_ids$self_report_non_cancer
#> code_fid date_fid 
#>  "20002"  "20008" 
#> 
#> $required_field_ids$self_report_non_cancer_icd10
#> code_fid date_fid 
#>  "20002"  "20008" 
#> 
#> $required_field_ids$self_report_cancer
#> code_fid date_fid 
#>  "20001"  "20006" 
#> 
#> $required_field_ids$self_report_operation
#> code_fid date_fid 
#>  "20004"  "20010" 
#> 
#> $required_field_ids$cancer_register_icd9
#> code_fid date_fid 
#>  "40013"  "40005" 
#> 
#> $required_field_ids$cancer_register_icd10
#> code_fid date_fid 
#>  "40006"  "40005" 
#> 
#> $required_field_ids$summary_hes_icd9
#> code_fid date_fid 
#>  "41271"  "41281" 
#> 
#> $required_field_ids$summary_hes_icd10
#> code_fid date_fid 
#>  "41270"  "41280" 
#> 
#> $required_field_ids$summary_hes_opcs3
#> code_fid date_fid 
#>  "41273"  "41283" 
#> 
#> $required_field_ids$summary_hes_opcs4
#> code_fid date_fid 
#>  "41272"  "41282" 
#> 
#>